requirements.txt
playwright
python-dotenv
colorama
requests